requirements.txt
blas==1.0
brotlipy==0.7.0
ca-certificates==2022.07.19
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
colorama==0.4.5
cpuonly==2.0
cryptography==37.0.1
dataclasses==0.8
flask==2.1.3
freetype==2.10.4
idna==3.4
importlib-metadata==4.11.3
intel-openmp==2021.4.0
itsdangerous==2.0.1
jinja2==3.0.3
joblib==1.2.0
jpeg==9e
lerc==3.0
libdeflate==1.8
libpng==1.6.37
libtiff==4.4.0
libuv==1.40.0
libwebp-base==1.2.4
libwebp==1.2.4
lz4-c==1.9.3
markupsafe==2.1.1
mkl-service==2.4.0
mkl==2021.4.0
mkl_fft==1.3.1
mkl_random==1.2.2
nltk==3.7
numpy-base==1.23.3
numpy==1.23.3
openssl==1.1.1q
pillow==9.2.0
pip==22.2.2
pycparser==2.21
pyopenssl==22.0.0
pysocks==1.7.1
python==3.8.13
pytorch-mutex==1.0
pytorch==1.12.1
regex==2022.9.13
requests==2.28.1
setuptools==63.4.1
six==1.16.0
sqlite==3.39.3
tk==8.6.12
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.11
vc==14.2
vs2015_runtime==14.27.29016
werkzeug==2.0.3
wheel==0.37.1
win_inet_pton==1.1.0
wincertstore==0.2
xz==5.2.6
zipp==3.8.0
zlib==1.2.12
zstd==1.5.2
gunicorn==20.1.0
